Split up Play view into multiple files
