+++ .vscode/settings.json
@@ -1,2 +1,4 @@
 {
+    "typescript.tsdk": "node_modules/typescript/lib",
+    "typescript.enablePromptUseWorkspaceTsdk": true,
 }
